fix: address PR comments (almost done)
